// src/types/Passenger.ts
export type Passenger = {
    firstName: string;
    lastName: string;
    gender: string;
    homeCountry: string;
    dateOfBirth: string;
    musterStatus: string;
    nationalIdNumber?: string;
    emergencyContactNumber?: string;
    specialCareOrAssistance?: string;
    customData?: Object;
}